requirements.txt
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
fasteners==0.19
Flask==3.0.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-apitools==0.5.31
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.11.3
google-cloud-bigquery-storage==2.22.0
google-cloud-bigtable==2.17.0
google-cloud-core==2.3.3
google-cloud-datastore==2.16.1
google-cloud-dlp==3.12.2
google-cloud-documentai==2.18.0
google-cloud-language==2.10.1
google-cloud-pubsub==2.17.1
google-cloud-pubsublite==1.8.3
google-cloud-recommendations-ai==0.10.4
google-cloud-secret-manager==2.17.0
google-cloud-spanner==3.36.0
google-cloud-storage==2.10.0
google-cloud-videointelligence==2.11.3
google-cloud-vision==3.4.4
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2
overrides==6.5.0
packaging==23.2
pandas==2.1.4
pip-tools==7.3.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.1
wheel==0.41.2
